Rename newUseApi to useApi
