+++ src/Hedwig/ClientApp/src/containers/Enrollment/Detail/EnrollmentDetail.test.tsx
@@ -3,10 +3,10 @@ import { mockAllFakeEnrollments, mockSite } from '../../../tests/data';
 import mockUseApi, {
 	mockApiOrganizationsOrgIdSitesSiteIdEnrollmentsIdGet,
 	mockApiOrganizationsOrgIdSitesIdGet,
-} from '../../../hooks/__mocks__/newUseApi';
+} from '../../../hooks/__mocks__/useApi';
 
 // Jest mocks must occur before later imports
-jest.mock('../../../hooks/newUseApi', () =>
+jest.mock('../../../hooks/useApi', () =>
 	mockUseApi({
 		apiOrganizationsOrgIdSitesIdGet: mockApiOrganizationsOrgIdSitesIdGet(mockSite),
 		apiOrganizationsOrgIdSitesSiteIdEnrollmentsIdGet: mockApiOrganizationsOrgIdSitesSiteIdEnrollmentsIdGet(
