+++ src/Hedwig/ClientApp/src/containers/Enrollment/_sections/FamilyInfo.tsx
@@ -21,7 +21,7 @@ import { validationErrorAlert } from '../../../utils/stringFormatters/alertTextM
 import AlertContext from '../../../contexts/Alert/AlertContext';
 import { FormReducer, formReducer, updateData } from '../../../utils/forms/form';
 import { DeepNonUndefineable } from '../../../utils/types';
-import useNewUseApi, { ApiError } from '../../../hooks/newUseApi';
+import useNewUseApi, { ApiError } from '../../../hooks/useApi';
 
 const FamilyInfo: Section = {
 	key: 'family-information',
