+++ src/Hedwig/ClientApp/src/containers/Reports/ReportDetail/ReportDetail.test.tsx
@@ -10,12 +10,12 @@ import {
 } from '../../../tests/data';
 import mockUseApi, {
 	mockApiOrganizationsOrgIdEnrollmentsGet,
-} from '../../../hooks/__mocks__/newUseApi';
+} from '../../../hooks/__mocks__/useApi';
 
 let mockReport = mockDefaultReport;
 
 // Jest mocks must occur before later imports
-jest.mock('../../../hooks/newUseApi', () =>
+jest.mock('../../../hooks/useApi', () =>
 	mockUseApi({
 		apiOrganizationsOrgIdEnrollmentsGet: mockApiOrganizationsOrgIdEnrollmentsGet([
 			mockCompleteEnrollment,
