+++ src/Hedwig/ClientApp/src/containers/Reports/ReportsSummary/ReportsSummary.tsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import { Table, TableProps } from '../../../components/Table/Table';
 import dateFormatter from '../../../utils/dateFormatter';
 import UserContext from '../../../contexts/User/UserContext';
-import newUseApi from '../../../hooks/newUseApi';
+import useApi from '../../../hooks/useApi';
 import {
 	ApiOrganizationsOrgIdReportsGetRequest,
 	CdcReport as Report,
@@ -18,14 +18,14 @@ export default function ReportsSummary() {
 	const orgParams: ApiOrganizationsIdGetRequest = {
 		id: getIdForUser(user, 'org'),
 	};
-	const { loading: orgLoading, error: orgError, data: organization } = newUseApi(
+	const { loading: orgLoading, error: orgError, data: organization } = useApi(
 		api => api.apiOrganizationsIdGet(orgParams),
 		{ skip: !user }
 	);
 	const reportParams: ApiOrganizationsOrgIdReportsGetRequest = {
 		orgId: getIdForUser(user, 'org'),
 	};
-	const { loading, error, data: reports } = newUseApi(
+	const { loading, error, data: reports } = useApi(
 		api => api.apiOrganizationsOrgIdReportsGet(reportParams),
 		{ skip: !user }
 	);
