+++ src/Hedwig/ClientApp/src/utils/validations/isBlockingValidationError.ts
@@ -1,4 +1,4 @@
-import { ApiError } from '../../hooks/newUseApi';
+import { ApiError } from '../../hooks/useApi';
 import { ValidationProblemDetails } from '../../generated';
 
 export function isBlockingValidationError(error: ApiError): boolean {
